kamehameha

+++ database/migrations/2024_12_04_072755_create_questions_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->string('type');
             $table->boolean('is_required');
             $table->text('correct_answer')->nullable();
-            $table->integer('points');
+            $table->float('points');
             $table->timestamps();
         });
     }
